pyproject.toml
[tool.poetry]
name = "bond-agent"
version = "3.0.0"
description = ""
authors = ["scc <scc@teamt5.org>"]
readme = "README.md"
packages = [{ include = "libs" }, { include = "static" }]

[tool.poetry.dependencies]
python = "^3.8,<3.13"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
loguru = "^0.7.0"
requests = "^2.31.0"
aiofiles = "^23.1.0"
anyio = "^3.7.0"
certifi = "^2023.5.7"
charset-normalizer = "^3.1.0"
click = "^8.1.3"
h11 = "^0.14.0"
idna = "^3.7"
pydantic = "^1.10.8"
python-multipart = "^0.0.7"
sniffio = "^1.3.0"
starlette = "^0.27.0"
urllib3 = "^2.0.2"
pillow = "^10.3.0"
werkzeug = "^2.3.4"

httpx = "^0.24.1"
pystray = { version = "^0.19.4", markers = "sys_platform == 'win32'" }
pyuac = { version = "^0.0.3", markers = "sys_platform == 'win32'" }
pywin32 = { version = "^306", markers = "sys_platform == 'win32'" }
jinja2 = "^3.1.4"
black = "^23.7.0"

psutil = { version = "^5.9.5", markers = "sys_platform == 'win32'" }
auto-py-to-exe = "^2.42.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
mock = "^5.0.2"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "./libs", "./static"]
